Adding InternerWithUint64Id type
The InternerWithUint64Id can be used to build a string interner for types
which can be uniquely identified by a single uint64 integer.

Three examples of types which can be identified by a uint64 value
include.

* int64 - This is easy the value itself clearly identifies itself after
  conversion to uint64
* float64 - We can use math.Float64bits(...) to get an identifying
  uint64
* time.Time - The time.UnixNanos() can uniquely identify a timestamp
  (although there is a problem with different timezones)

The basic principle here is that in order to intern a string value we
need to correctly identify the value being converted into a string, and
then look for it in our set of interned strings. If the value already
has an interned string then we don't need to build and allocate a new
string, we just return the existing interned string.

There are two parts to the InternerWithUint64.

1: Converter

type ConverterWithUint64Id interface {
        Identity() uint64
        String() string
}

This is a basic interface which wraps the type to be converted to a
string. We need to provide a uint64 value to identify the value uniquely
and also a string value, if we need to generate a string representation
of this value.

2: Interner

type InternerWithUint64Id[C ConverterWithUint64Id] struct {
	...
}

func (i *InternerWithUint64Id[C]) Get(converter C) string {
	...
}

This struct actually does the work of interning string representations.
It will look up the identity of the value, and if it exists already
return the previously interned string. Otherwise if will call the
String() method on the converter to get the string representation and
then intern that string value for future use.

We can see that the interner's type is parameterised by a specific
converter.

An implementation detail is that the interner will run a number of
interal shards in order to increase the concurrency available for
interning.