Adding InternerWithBytesId type
The InternerWithBytesId can be used to build a string interner for types
which can be uniquely identified by a []byte.

Two examples of types which can be identified by a []byte are

* []byte - This is very easy
* string - Also very easy, note because we never modify the []byte here
  we can unsafely convert a string -> []byte and avoid an allocation.

The Converter interface looks like

type ConverterWithBytesId interface {
        Identity() []byte
}

Here we only require that the converter return the identifying []byte.
This is because we _require_ that the []byte representation is
byte-for-byte the same as the string representation. So we the
InternerWithBytesId performs this conversion internally to preserve this
property.

The interner interface looks like

type InternerWithBytesId[C ConverterWithBytesId] struct {
	...
}

func (i *InternerWithBytesId[C]) Get(converter C) string {
	...
}

This works basically the same as the InternerWithInt64Id, but curiously
internally the []byte is hashed into a uint64 value. So values are
_still_ effectively identified by a uint64. However, when we retrieve an
interned string we check to see if the interned string exactly matches
the []byte representation, so the identity is ultimately determined by
the full value of the []byte. In the case where the hash value retrieves
a string which doesn't match the []byte identity the []byte cannot be
interned (ever) and is recorded as a hash-collision in the stats. This
is probably very very rare.