Reorganise project - offheap at root
This project is now all about the offheap package. It sits at the root
of the project.

Inside pkg/ we have packages which are built using the offheap package.
Currently quadtree, intern, linkedlist.

Inside testpkg/ we have packages which are test utilities. Currently
fuzzutil and testutil.

All remaining code has been removed and is not longer a meaningful part
of this project.

rename imports