Add a range of benchmarks for interners
We benchmark the following scenarios.

1: Get(...) b.N many new values. Expected to be slow and to allocate
   memory.

2: Get(...) b.N already interned values. Expected to be fast and to
   _never_ allocate.

3: Get(...) 10,000 already interned values b.N times. This is the same
   as scenario 2 above, but simulates the performance when the set of
   distinct values is relatively small and repeated. Expected to be fast
   and to _never_ allocate.

It is worth noting that all of the tests and all of the benchmarks are
copy/paste jobs of each other. I did attempt to build a more generic
version of these tests, to avoid the copy and pasting, but the result
was verbose confusing to read and _also_ required a lot of copy/paste to
manage.

The frankly extreme level of copy/paste in these tests is pointing to
some design or testing problem. We need to rethink this in the future.
It may be the case that it's primarily an API design issue.