Switch interners over to Interner[T any] interface
We have introduced a new interface

type Interner[T any] interface {
	Get(t T) string
	GetStats() StatsSummary
}

This is now the type of every implemented interner.

The concrete type of each interner are no longer exported. So the only
way to build these types is through their exported constructors.

There's no necessity for hiding the concrete types of each interner, but
I feel like because this package is intended to also support building
interners it has a lot of types. This can get confusing very quickly. So
keeping it to the absolute minimum is valuable.