Creating internbase package
This allows the types and structs which are used to build each interner
to live in a separate package. This means that first time users of the
intern package should be able to see the actual interners directly and
immediately. Previously there were more types which we used to build
interners than actual interners themselves. This would make using the
package very challenging when first encountering the package.