Unexport all Converter structs
These types were all exported and public. But there is no real benefit
to this. It makes it look as though people using the intern package
should expect to directly user or instantiate one of these converters.
This seems very unlikely to be the case in practice and they are a very
hidable implementation detail.

This should really clean up the godoc for this package.