Adding gcassert:noescape to all RefPointer methods
Each of these methods have a pointer receiver. But RefPointer is used as
a value, so it's important that the receiver does not escape to the
heap. If it does escape then each method call(!) will allocate to the
heap.

As a side note, the rationale behind using a pointer receiver is to
avoid copying two words on RefPointer method calls. This rationale makes
sense, but hasn't been tested as a performance improvement. In the
future we should test this. Maybe value receivers are just as good (or
better).