metadata now has dataAddressAndGen field
This replaces the old 'gen uint8' we had before. The behaviour does not
change in this commit, we are simply storing the generation tag in a
larger field than before. The next steps will be to actually use this
new full address to do the data lookup.

Also added a helper function to test the validity of a
metadata/reference pair. This functionality was being duplicated between
Free() and DataPtr() methods.