checkReference() checks only the raw value
Previously we separately checked that the generation tags were the same.
But this is redundant. Now we just check that the raw address values are
the same and print an explanatory error in the panic including the
taggedAddress values as hex. This should be enough to debug and error in
this case.

Just printing big raw hex values _could_ be harder to debug for users
who are not already familiar with the internals of the offheap package.
Maybe a more beginner friendly message could be used instead.