Replace linked-list of free objects with slice
This is a much simpler implementation. It also removed object meta-data
from the critical path of tracking free objects.

Now meta-data _only_ checks if the object is free or if the object has
the right generation tag for the accessing reference. These functions
are useful, but not essential, so the meta-data can be made optional.