Add isFree bit to taggedAddress
This shrinks the size of the generation tag to 7 bits. I think this is
likely fine in practice. The new maximum generation size is 127 or 0x7F
in hexadecimal.

You can set and read the is-free bit with specific methods on
taggedAddress.

This is not currently used. Will be actually used in the next commit.