RefPointer _only_ contains address of metadata
The metadata now contains the address of the actual data. This means
that the RefPointer is only 64 bits wide. This also means that right now
you must do two dependent memory lookups to get to the data. This is
_probably_ slower than previously where we also did two memory lookups
but they were independent and speculative execuation _probably_
initiates the second lookup earlier.

This slow down is ok (and only theoretical, not benchmarked or
conclusively tested).

Making the RefPointer only 64 bits wide allows us to do vanilla atomic
operations like CAS etc. Previously doing CAS on two words looked
potentially very complex and likely much slower.

We will introduce a const flag in the future which can be switched off
at build time to bypass the metadata lookup and store the data address
directly in the RefPointer. This will speed up data access times, but
only when the flag is switched off at compile time. A good idea? A bad
idea? Ony time will tell. (It's for sure an idea)