Check RefPointer is valid on Realloc
Previously we weren't checking this, so any old Reference could
Realloc() an allocation at any time, even if that Reference was no
longer valid.

It's concerning that all this work has uncovered multiple holes in the
Reference generation checking logic. It's not clear how to improve this
beyond just thinking harder and reviewing code more carefully (neither
of these approaches are very reliable ways to find bugs in general). But
we will keep this problem in the back of our heads and see if we can
improve things.