Rename dataAddress() -> accessibleActiveAddress()
This RefPointer method does all the generation/is-free checks on an
allocation. It's important that this method sits on the path of any
RefPointer method which accesses the data or meta-data.

Current methods which call this method are:

Realloc()
DataPtr()
Bytes()
free()

Also inlined the metadata.checkReference() method.  The check in
metadata _only_ checked that the generation matches, which doesn't
really warrant a separate method. The code seems more readable with the
generation check inlined into the accessibleActiveAddress() method.