Remove pointer receivers from RefPointer methods
Because the RefPointer is only one word wide, and no methods update the
receiver, they return new modified RefPointers instead, we don't need
pointer receiver semantics.

This is a significant simplification of this part of the code-base.
Previously all RefPointer, and all Ref-types in general, are used as
values but have methods with pointer receivers. This caused some
unexpected allocations which could be fixed, but with odd-looking
contortions in the code and long comments to explain them.

Now there are no pointer receivers, no subtle allocations. Happy day.