Update metadata comments
Clarify that metadata contains both the generation _and_ the is-free
tag, which are used to check the integrity of RefPointer data accesses.

Writing these comments it strikes me that metadata is responsible for
the integrity checks when accessing data through Refpointers, but the
check themselves are done inside RefPointer methods. It seems like
moving this code into metadata will improve code clarity.