Adding taggedAddress.withIncGen() method
This method simply creates a new taggedAddress with an incremented
generation tag. Only metadata increments the generation tag, while on
RefPointer the generation tag is set directly. This is because when we
allocFromFree a RefPointer the generation tag could be any legal value.

This is not a critical change, but tidies up a little bit of code.
Hopefully it is easier to read now.