containsNoPointer() func takes reflect.Type arg
Previously we passed the type in via a generic parameter. Passing it in
as a struct allows us to build the reflect type and cache it outside of
the containsNoPointer() function.

The store itself will cache the results to this function call to avoid
calling it every time we allocate a type.