Add cache for type check
Previously every allocation would perform a check on whether the
allocation type contains pointers. This work is very slow and allocates
a remarkable amount of memory.

We introduce a cache which simply stores types which have been
successfully checked previously. This cache is protected by a read/write
lock - which does put another lock on the allocation path.