Merge pointer_checker.go into type_checker.go
There is only one file which contains all of the type checking code in
it.

We also add tests which directly test the pointerCheck type as well as
the containsNoPointers() function. We preserve the lower level function
tests because they are able to verify that the exact type failures we
expect are being found, and described in the error text.